fixes music sound to some degree - fmv sound breaks as the format is hard coded to float till music is fixed

+++ include/oddlib/audio/Sample.h
@@ -8,5 +8,5 @@ public:
     Uint16 * m_SampleBuffer;
     unsigned int i_SampleSize;
 
-    Uint16 GetSample(Uint32 sampleOffset, bool interpolation);
+    float GetSample(float sampleOffset, bool interpolation);
 };
